Moved Class Form into Own Component to prepare for edit mode
