+++ client/src/components/ClassForm/ImageUpload.scss
