+++ client/src/constants/ClassCategories.js
